move EntityAndRepository mapper interface to Doctrine package, require min Symfony 4.4
