+++ composer.json
@@ -12,35 +12,35 @@
     "require": {
         "php": "^7.1",
         "ext-json": "*",
-        "composer/xdebug-handler": "^1.3",
-        "doctrine/annotations": "^1.6",
+        "composer/xdebug-handler": "^1.4",
+        "doctrine/annotations": "^1.8",
         "doctrine/inflector": "^1.3",
         "jean85/pretty-package-versions": "^1.2",
-        "nette/robot-loader": "^3.1",
-        "nette/utils": "^2.5|^3.0",
+        "nette/robot-loader": "^3.2",
+        "nette/utils": "^3.0",
         "nikic/php-parser": "^4.3",
         "ondram/ci-detector": "^3.1",
         "phpstan/phpdoc-parser": "^0.3.5",
-        "phpstan/phpstan": "^0.11.13",
+        "phpstan/phpstan": "^0.11.19",
         "phpstan/phpstan-phpunit": "^0.11.2",
         "sebastian/diff": "^3.0",
-        "symfony/console": "^3.4|^4.4|^5.0",
-        "symfony/dependency-injection": "^3.4|^4.4|^5.0",
-        "symfony/finder": "^3.4|^4.4|^5.0",
-        "symfony/process": "^3.4|^4.4|^5.0",
+        "symfony/console": "^4.4|^5.0",
+        "symfony/dependency-injection": "^4.4|^5.0",
+        "symfony/finder": "^4.4|^5.0",
+        "symfony/process": "^4.4|^5.0",
         "symplify/package-builder": "^6.1"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.14",
-        "jetbrains/phpstorm-stubs": "^2019.1",
-        "ocramius/package-versions": "^1.4",
-        "phpunit/phpunit": "^7.5|^8.0",
+        "friendsofphp/php-cs-fixer": "^2.16",
+        "jetbrains/phpstorm-stubs": "^2019.2",
+        "ocramius/package-versions": "^1.5",
+        "phpunit/phpunit": "^7.5|^8.4",
         "symplify/changelog-linker": "^6.1",
         "symplify/easy-coding-standard": "^6.1",
         "symplify/monorepo-builder": "^6.1",
         "symplify/phpstan-extensions": "^6.1",
-        "thecodingmachine/phpstan-strict-rules": "^0.11",
-        "tracy/tracy": "^2.5"
+        "thecodingmachine/phpstan-strict-rules": "^0.11.2",
+        "tracy/tracy": "^2.7"
     },
     "autoload": {
         "psr-4": {
