+++ packages/Doctrine/src/Contract/Mapper/DoctrineEntityAndRepositoryMapperInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Bridge\Contract;
+namespace Rector\Doctrine\Contract\Mapper;
 
 interface DoctrineEntityAndRepositoryMapperInterface
 {
